+++ pkg/dbt_cloud/spark_credential.go
@@ -26,7 +26,7 @@ type SparkUnencryptedCredentialDetails struct {
 
 type SparkCredential struct {
 	ID                           *int                              `json:"id"`
-	Account_Id                   int                               `json:"account_id"`
+	Account_Id                   int64                              `json:"account_id"`
 	Project_Id                   int                               `json:"project_id"`
 	Type                         string                            `json:"type"`
 	State                        int                               `json:"state"`
@@ -39,7 +39,7 @@ type SparkCredential struct {
 
 type SparkCredentialGlobConn struct {
 	ID                *int                     `json:"id"`
-	AccountID         int                      `json:"account_id"`
+	AccountID         int64                    `json:"account_id"`
 	ProjectID         int                      `json:"project_id"`
 	Type              string                   `json:"type"`
 	State             int                      `json:"state"`
